# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HydroHeat AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface City {
  id: string;
  name: string;
  state: string;
  lat: number;
  lng: number;
  temperature: number;
  groundwater: number;
  riskScore: number;
  population: number;
  heatIslandIntensity: number;
  airQualityIndex: number;
  humidity: number;
  rainfall: number;
}

export type InterventionType =
  (typeof InterventionType)[keyof typeof InterventionType];

export const InterventionType = {
  tree: "tree",
  fountain: "fountain",
  solar: "solar",
  green_roof: "green_roof",
  permeable_pavement: "permeable_pavement",
  factory: "factory",
  stubble_burning: "stubble_burning",
  fireworks: "fireworks",
} as const;

export interface Intervention {
  type: InterventionType;
  lat: number;
  lng: number;
  count?: number;
}

/**
 * Optional geo context to make physics-based adjustments
 */
export interface GeoSimulationContext {
  ndvi: number;
  waterBodyDistanceKm: number;
  landUseClass: string;
  soilMoisturePercent: number;
  urbanDensityPercent: number;
  treeEffectivenessMultiplier: number;
  permeablePavementEffectiveness: number;
  groundwaterRechargePotential: number;
}

export interface SimulationRequest {
  cityId: string;
  interventions: Intervention[];
  geoContext?: GeoSimulationContext;
}

export interface InterventionEffect {
  type: string;
  effect: string;
  temperatureImpact: number;
  groundwaterImpact: number;
}

export interface SimulationResult {
  cityId: string;
  temperatureDelta: number;
  groundwaterDelta: number;
  riskScoreDelta: number;
  airQualityDelta: number;
  projectedTemperature: number;
  projectedGroundwater: number;
  projectedRiskScore: number;
  aiInsights: string[];
  interventionBreakdown: InterventionEffect[];
  /** Whether geo context was applied to adjust physics */
  geoAdjusted: boolean;
}

export interface WaterBodyInfo {
  name: string;
  type: string;
  distanceKm: number;
  influenceLevel: string;
}

export interface RealWeatherData {
  temperature: number;
  humidity: number;
  windSpeed: number;
  precipitation: number;
  cloudCover: number;
  soilMoisturePercent: number;
  evapotranspiration: number;
  source: string;
}

/**
 * ISRO Bhuvan-style satellite data analysis
 */
export interface ISROAnalysis {
  lulcClass: string;
  lulcSubClass: string;
  vegetationHealthIndex: number;
  surfaceAlbedo: number;
  evapotranspirationRate: number;
  landsatBand: string;
  carbonSequestrationKgPerHa: number;
  soilErosionRisk: string;
  floodRisk: string;
  heatIslandCategory: string;
}

/**
 * Physics multipliers derived from geo context
 */
export interface SimulationModifiers {
  treeEffectivenessMultiplier: number;
  permeablePavementEffectiveness: number;
  groundwaterRechargePotential: number;
  heatIslandSeverity: number;
  solarEffectiveness: number;
  factoryImpactMultiplier: number;
}

export interface GeoAnalysis {
  lat: number;
  lng: number;
  /** Normalized Difference Vegetation Index (-1 to 1) */
  ndvi: number;
  /** Vegetation classification */
  ndviClass: string;
  /** Vegetation health 0-100 */
  ndviHealthScore: number;
  nearestWaterBody?: WaterBodyInfo;
  /** Number of water bodies within 100km */
  waterBodyCount: number;
  /** ISRO LULC classification */
  landUseClass: string;
  urbanDensityPercent: number;
  soilType: string;
  /** 0-100, higher = more drought-prone */
  droughtIndex: number;
  realWeather: RealWeatherData;
  isroAnalysis: ISROAnalysis;
  simulationModifiers: SimulationModifiers;
  geoRecommendations: string[];
  /** Data provenance note */
  dataSource: string;
}

export interface WeatherData {
  lat: number;
  lng: number;
  temperature: number;
  humidity: number;
  windSpeed: number;
  windDirection: string;
  conditions: string;
  uvIndex: number;
  feelsLike: number;
  visibility: number;
}