# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HydroHeat AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /geo-analysis/{lat}/{lng}:
    get:
      operationId: getGeoAnalysis
      tags: [geo]
      summary: Get geo-spatial intelligence for coordinates
      description: >
        Returns ISRO-style satellite analysis including NDVI vegetation index,
        water body detection, LULC classification, real weather data from Open-Meteo,
        and geo-contextual simulation modifiers.
      parameters:
        - name: lat
          in: path
          required: true
          schema:
            type: number
        - name: lng
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Geo-spatial analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeoAnalysis"
...
components:
  schemas:
...
    GeoAnalysis:
      type: object
      properties:
        lat:
          type: number
        lng:
          type: number
        ndvi:
          type: number
          description: Normalized Difference Vegetation Index (-1 to 1)
        ndviClass:
          type: string
          description: Vegetation classification
        ndviHealthScore:
          type: number
          description: Vegetation health 0-100
        nearestWaterBody:
          $ref: "#/components/schemas/WaterBodyInfo"
        waterBodyCount:
          type: integer
          description: Number of water bodies within 100km
        landUseClass:
          type: string
          description: ISRO LULC classification
        urbanDensityPercent:
          type: number
        soilType:
          type: string
        droughtIndex:
          type: number
          description: 0-100, higher = more drought-prone
        realWeather:
          $ref: "#/components/schemas/RealWeatherData"
        isroAnalysis:
          $ref: "#/components/schemas/ISROAnalysis"
        simulationModifiers:
          $ref: "#/components/schemas/SimulationModifiers"
        geoRecommendations:
          type: array
          items:
            type: string
        dataSource:
          type: string
          description: Data provenance note
      required:
        - lat
        - lng
        - ndvi
        - ndviClass
        - ndviHealthScore
        - waterBodyCount
        - landUseClass
        - urbanDensityPercent
        - soilType
        - droughtIndex
        - realWeather
        - isroAnalysis
        - simulationModifiers
        - geoRecommendations
        - dataSource
    WaterBodyInfo:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        distanceKm:
          type: number
        influenceLevel:
          type: string
      required:
        - name
        - type
        - distanceKm
        - influenceLevel
    RealWeatherData:
      type: object
      properties:
        temperature:
          type: number
        humidity:
          type: number
        windSpeed:
          type: number
        precipitation:
          type: number
        cloudCover:
          type: number
        soilMoisturePercent:
          type: number
        evapotranspiration:
          type: number
        source:
          type: string
      required:
        - temperature
        - humidity
        - windSpeed
        - precipitation
        - cloudCover
        - soilMoisturePercent
        - evapotranspiration
        - source
    ISROAnalysis:
      type: object
      description: ISRO Bhuvan-style satellite data analysis
      properties:
        lulcClass:
          type: string
        lulcSubClass:
          type: string
        vegetationHealthIndex:
          type: number
        surfaceAlbedo:
          type: number
        evapotranspirationRate:
          type: number
        landsatBand:
          type: string
        carbonSequestrationKgPerHa:
          type: number
        soilErosionRisk:
          type: string
        floodRisk:
          type: string
        heatIslandCategory:
          type: string
      required:
        - lulcClass
        - lulcSubClass
        - vegetationHealthIndex
        - surfaceAlbedo
        - evapotranspirationRate
        - landsatBand
        - carbonSequestrationKgPerHa
        - soilErosionRisk
        - floodRisk
        - heatIslandCategory
    SimulationModifiers:
      type: object
      description: Physics multipliers derived from geo context
      properties:
        treeEffectivenessMultiplier:
          type: number
        permeablePavementEffectiveness:
          type: number
        groundwaterRechargePotential:
          type: number
        heatIslandSeverity:
          type: number
        solarEffectiveness:
          type: number
        factoryImpactMultiplier:
          type: number
      required:
        - treeEffectivenessMultiplier
        - permeablePavementEffectiveness
        - groundwaterRechargePotential
        - heatIslandSeverity
        - solarEffectiveness
        - factoryImpactMultiplier
...