# lib/api-zod/src/generated/types/geoAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HydroHeat AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ISROAnalysis } from "./iSROAnalysis";
import type { RealWeatherData } from "./realWeatherData";
import type { SimulationModifiers } from "./simulationModifiers";
import type { WaterBodyInfo } from "./waterBodyInfo";

export interface GeoAnalysis {
  lat: number;
  lng: number;
  /** Normalized Difference Vegetation Index (-1 to 1) */
  ndvi: number;
  /** Vegetation classification */
  ndviClass: string;
  /** Vegetation health 0-100 */
  ndviHealthScore: number;
  nearestWaterBody?: WaterBodyInfo;
  /** Number of water bodies within 100km */
  waterBodyCount: number;
  /** ISRO LULC classification */
  landUseClass: string;
  urbanDensityPercent: number;
  soilType: string;
  /** 0-100, higher = more drought-prone */
  droughtIndex: number;
  realWeather: RealWeatherData;
  isroAnalysis: ISROAnalysis;
  simulationModifiers: SimulationModifiers;
  geoRecommendations: string[];
  /** Data provenance note */
  dataSource: string;
}