# lib/api-zod/src/generated/types/iSROAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HydroHeat AI API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * ISRO Bhuvan-style satellite data analysis
 */
export interface ISROAnalysis {
  lulcClass: string;
  lulcSubClass: string;
  vegetationHealthIndex: number;
  surfaceAlbedo: number;
  evapotranspirationRate: number;
  landsatBand: string;
  carbonSequestrationKgPerHa: number;
  soilErosionRisk: string;
  floodRisk: string;
  heatIslandCategory: string;
}