# lib/api-zod/src/generated/types/realWeatherData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HydroHeat AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RealWeatherData {
  temperature: number;
  humidity: number;
  windSpeed: number;
  precipitation: number;
  cloudCover: number;
  soilMoisturePercent: number;
  evapotranspiration: number;
  source: string;
}