# lib/api-zod/src/generated/types/simulationModifiers.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HydroHeat AI API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Physics multipliers derived from geo context
 */
export interface SimulationModifiers {
  treeEffectivenessMultiplier: number;
  permeablePavementEffectiveness: number;
  groundwaterRechargePotential: number;
  heatIslandSeverity: number;
  solarEffectiveness: number;
  factoryImpactMultiplier: number;
}