# lib/api-zod/src/generated/types/waterBodyInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HydroHeat AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WaterBodyInfo {
  name: string;
  type: string;
  distanceKm: number;
  influenceLevel: string;
}